코드 리뷰 반영 : SplitUtils 생성 (문자열 분리 클래스 분리)
