+++ src/main/java/Calculator/StringAddCalculator.java
@@ -2,19 +2,11 @@ package Calculator;
 
 import java.util.Arrays;
 import java.util.Objects;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class StringAddCalculator {
 
-    private static final String REGEX_DELIMITER_DEFAULT = "[,:]";
-    private static final Pattern REGEX_DELIMITER_PATTERN = Pattern.compile("//(.)\\n(.*)");
-
     private static final int ZERO = 0;
     private static final String EMPTY_VALUE = "";
-
-    private static final int MATCHER_DELIMITER_GROUP = 1;
-    private static final int MATCHER_EXPRESSION_GROUP = 2;
 
     private static final String MESSAGE_INVALID_INPUT = "입력값은 0이상의 숫자여야 합니다.";
     private static final String MESSAGE_UTILITY_CLASS = "Utility class";
@@ -28,24 +20,11 @@ public class StringAddCalculator {
             return ZERO;
         }
 
-        return sum(split(input));
+        return sum(SplitUtils.split(input));
     }
 
     private static boolean isValidInput(String input) {
         return Objects.isNull(input) || input.equals(EMPTY_VALUE);
-    }
-
-    private static String[] split(String input) {
-        Matcher matcher = REGEX_DELIMITER_PATTERN.matcher(input);
-
-        if (matcher.find()) {
-            return split(matcher.group(MATCHER_DELIMITER_GROUP), matcher.group(MATCHER_EXPRESSION_GROUP));
-        }
-        return split(REGEX_DELIMITER_DEFAULT, input);
-    }
-
-    private static String[] split(String delimiter, String expression) {
-        return expression.split(delimiter);
     }
 
     private static int sum(String[] list) {
